+++ src/server/IServerAppConfig.ts
@@ -32,6 +32,13 @@ export interface IServerAppConfig {
   json?: boolean
   log?: boolean
   session?: boolean
+  sessionCookieKey?: string
+  sessionHttpOnly?: boolean
+  sessionMaxAge?: number | 'session'
+  sessionOverwrite?: boolean
+  sessionRenew?: boolean
+  sessionRolling?: boolean
+  sessionSigned?: boolean
 
   middleware?: Koa.Middleware[]
 }